Moving common code out of fluvio-cli
