+++ crates/fluvio-cli/Cargo.toml
@@ -81,6 +81,7 @@ fluvio-extension-common = { path = "../fluvio-extension-common", features = ["ta
 fluvio-controlplane-metadata = { path = "../fluvio-controlplane-metadata", features = ["use_serde", "k8"] }
 k8-types = { version = "0.5.0", features = ["core"] }
 fluvio-channel = { path = "../fluvio-channel" }
+fluvio-cli-common = { path = "../fluvio-cli-common" }
 
 # Optional Fluvio dependencies
 fluvio-types = { path = "../fluvio-types", optional = true }
